+++ demo/web/typescript/index.html
@@ -3,12 +3,12 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>rclnodejs/web — end-to-end typed in a TypeScript app</title>
+    <title>rclnodejs/web — end-to-end typed ROS 2 in a TypeScript app</title>
     <link rel="stylesheet" href="/src/style.css" />
   </head>
   <body>
     <h1>
-      rclnodejs/web demo
+      End-to-end typed ROS 2 in a TypeScript app
       <span class="badge">TypeScript</span>
     </h1>
     <p>
